feat: Move `NoteBuilder` to `miden-lib`
