+++ crates/miden-lib/src/testing/mod.rs
@@ -1,3 +1,4 @@
 pub mod account_component;
 pub mod mock_account;
 pub mod mock_account_code;
+pub mod note;
