+++ crates/miden-testing/src/kernel_tests/batch/proposed_batch.rs
@@ -4,6 +4,7 @@ use std::collections::BTreeMap;
 use anyhow::Context;
 use assert_matches::assert_matches;
 use miden_lib::testing::account_component::MockAccountComponent;
+use miden_lib::testing::note::NoteBuilder;
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::account::{Account, AccountId, AccountStorageMode};
 use miden_objects::batch::ProposedBatch;
@@ -11,7 +12,6 @@ use miden_objects::block::BlockNumber;
 use miden_objects::crypto::merkle::MerkleError;
 use miden_objects::note::{Note, NoteType};
 use miden_objects::testing::account_id::AccountIdBuilder;
-use miden_objects::testing::note::NoteBuilder;
 use miden_objects::transaction::{InputNote, InputNoteCommitment, OutputNote, PartialBlockchain};
 use miden_objects::{BatchAccountUpdateError, ProposedBatchError, Word};
 use rand::rngs::SmallRng;
